requirements/develop.txt
asgiref==3.8.1
black @ git+https://github.com/psf/black@3be19b306f77980349d2c9edece2d86b1f58663d
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
factory_boy==3.3.0
filelock==3.14.0
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==22.0.0
identify==2.5.36
idna==3.7
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pep8-naming==0.13.3
phonenumberslite==8.13.37
platformdirs==4.2.1
pre-commit==3.7.1
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
setuptools==70.1.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
